pc-receiver/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyopenssl==24.2.1
qrcode==7.4.2
Pillow==10.3.0